перенес обращение к Api в основной компонент
